refactor: NewDriftCTL interface
