+++ pkg/driftctl.go
@@ -5,13 +5,28 @@ import (
 
 	"github.com/cloudskiff/driftctl/pkg/alerter"
 	"github.com/cloudskiff/driftctl/pkg/analyser"
+	"github.com/cloudskiff/driftctl/pkg/cmd/scan/output"
 	"github.com/cloudskiff/driftctl/pkg/filter"
+	"github.com/cloudskiff/driftctl/pkg/iac/config"
+	"github.com/cloudskiff/driftctl/pkg/iac/terraform/state/backend"
 	"github.com/cloudskiff/driftctl/pkg/middlewares"
 	"github.com/cloudskiff/driftctl/pkg/resource"
 	"github.com/jmespath/go-jmespath"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
+
+type ScanOptions struct {
+	Coverage       bool
+	Detect         bool
+	From           []config.SupplierConfig
+	To             string
+	Output         output.OutputConfig
+	Filter         *jmespath.JMESPath
+	Quiet          bool
+	BackendOptions *backend.Options
+	StrictMode     bool
+}
 
 type DriftCTL struct {
 	remoteSupplier resource.Supplier
@@ -22,8 +37,15 @@ type DriftCTL struct {
 	strictMode     bool
 }
 
-func NewDriftCTL(remoteSupplier resource.Supplier, iacSupplier resource.Supplier, filter *jmespath.JMESPath, alerter *alerter.Alerter, strictMode bool) *DriftCTL {
-	return &DriftCTL{remoteSupplier, iacSupplier, alerter, analyser.NewAnalyzer(alerter), filter, strictMode}
+func NewDriftCTL(remoteSupplier resource.Supplier, iacSupplier resource.Supplier, alerter *alerter.Alerter, opts *ScanOptions) *DriftCTL {
+	return &DriftCTL{
+		remoteSupplier,
+		iacSupplier,
+		alerter,
+		analyser.NewAnalyzer(alerter),
+		opts.Filter,
+		opts.StrictMode,
+	}
 }
 
 func (d DriftCTL) Run() (*analyser.Analysis, error) {
